New module: Gun
Added the beginnings of a "Gun" scene. Could use a custom GunStatblock: Resource to generate a bunch of different guns that could simply be fired with a button, regardless of what the effect will be. Currently the Gun just prints "pew!" to the console when fired. It's intended to be attached as a child to any willing body.